connection to postgresql fixed

+++ src/main/java/es/nomann/tddproject/data.sql
@@ -1,11 +1,11 @@
 CREATE TABLE city (
-    id INT  PRIMARY KEY,
+    id BIGINT  PRIMARY KEY,
     name VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE person (
-    id INT PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     username VARCHAR(255) NOT NULL,
     email VARCHAR(255) NOT NULL,
-    city_id INT REFERENCES city(id)
+    city_id BIGINT REFERENCES city(id)
 );
